+++ classes/BlueChip/Cache/Feeder.php
@@ -108,40 +108,23 @@ class Feeder
     /**
      * Get list of URLs to crawl.
      *
-     * This method fetches URLs from core XML sitemap providers, but this handling can be shortcut
-     * via `bc-cache/filter:cache-warm-up-initial-url-list` filter.
-     *
      * @return string[] List of URLs to crawl
      */
     private function getUrls(): array
     {
+        // Following filter allows to shortcut reading of XML sitemaps.
         $urls = apply_filters(Hooks::FILTER_CACHE_WARM_UP_INITIAL_URL_LIST, null);
 
         if ($urls === null) {
-            // If no URLs are provided by other means, use XML sitemaps providers from WordPress core.
-
-            /** @var \WP_Sitemaps_Provider[] $sitemap_providers */
-            $sitemap_providers = wp_get_sitemap_providers();
-
-            $url_sets = [];
-
-            foreach ($sitemap_providers as $sitemap_provider) {
-                foreach (\array_keys($sitemap_provider->get_object_subtypes()) as $object_subtype) {
-                    $max_num_pages = $sitemap_provider->get_max_num_pages($object_subtype);
-                    for ($page_num = 1; $page_num <= $max_num_pages; ++$page_num) {
-                        $url_sets[] = \array_column(
-                            $sitemap_provider->get_url_list($page_num, $object_subtype),
-                            'loc'
-                        );
-                    }
-                }
+            try {
+                $urls = XmlSitemapReader::getUrls();
+            } catch (Exception $e) {
+                // Trigger a warning and let WordPress handle it.
+                \trigger_error($e, E_USER_WARNING);
+                // Sorry, no URLs available.
+                $urls = [];
             }
-
-            $urls = \array_merge(...$url_sets);
         }
-
-        // Prepend home URL to the merged list of all URLs as it arguably represents the most important page on website.
-        \array_unshift($urls, home_url('/'));
 
         // Make sure only unique URLs are returned.
         return \array_unique(apply_filters(Hooks::FILTER_CACHE_WARM_UP_FINAL_URL_LIST, $urls));